libsel4allocman: overflow guards for large UT
seL4 can and does produce Untypeds that reach the end of the address
space. Using start + BIT(size) is not overflow safe for these.

Replace this pattern with MASK instead for inclusive-end reasoning, or
explicitly use uint64_t where that is too inconvenient.

The latter does rely on the fact that we will not get Untypeds that
reach 2^64. It does look like no architecture so far has a full 64-bit
physical address space, so this should be impossible.

This is not a full overflow safety survey of all of libsel4allocman,
just what I found by searching for "BIT(" in the untyped allocation
path. It at least does fix the errors I was getting for zcu102 which
does now produce such an untyped cap.

+++ libsel4allocman/src/bootstrap.c
@@ -1067,7 +1067,7 @@ static int handle_device_untyped_cap(add_untypeds_state_t *state, uintptr_t padd
 {
     bool cap_tainted = false;
     int error;
-    uintptr_t ut_end = paddr + BIT(size_bits);
+    uint64_t ut_end = (uint64_t)paddr + BIT(size_bits);
     int num_regions = 0;
     if (state != NULL) {
         num_regions = state->num_regions;
